Update plotme, gridme, add allcomb
plotme and gridme now specify the plotting dimension instead of using interaction indices. I will need to create a gridding function that creates a proper grid over which to simulate. allcomb has been added to help with this. Simulation plots need to be updated to switch from interaction indices to specifying plot_Xdim